// web/app/brief/evidence.generated.ts
export const CYCLES = [
  {
    "cycle": 1,
    "tokenId": "7268711",
    "mintIso": "2026-08-28T07:34:41.000Z",
    "closeIso": "2026-08-28T18:39:07.000Z",
    "heldSeconds": 39866,
    "entryNotionalUsd": 39.958556,
    "bnbMovePercent": -3.4555,
    "impermanentLossUsd": -0.126924,
    "swapFeesUsd": 0.017567,
    "cakeUsd": 0.002436,
    "gasUsd": 0.063557,
    "netVsHoldingUsd": -0.170478,
    "beatHolding": false
  },
  {
    "cycle": 2,
    "tokenId": "7272152",
    "mintIso": "2026-08-28T18:39:19.000Z",
    "closeIso": "2026-09-03T10:56:32.000Z",
    "heldSeconds": 490633,
    "entryNotionalUsd": 38.958173,
    "bnbMovePercent": 2.8031,
    "impermanentLossUsd": -0.081373,
    "swapFeesUsd": 0.095367,
    "cakeUsd": 0.039786,
    "gasUsd": 0.05884,
    "netVsHoldingUsd": -0.00506,
    "beatHolding": false
  },
  {
    "cycle": 3,
    "tokenId": "7316556",
    "mintIso": "2026-09-03T10:56:40.000Z",
    "closeIso": "2026-09-03T16:00:17.000Z",
    "heldSeconds": 18217,
    "entryNotionalUsd": 39.293144,
    "bnbMovePercent": 3.1883,
    "impermanentLossUsd": -0.10063,
    "swapFeesUsd": 0.01446,
    "cakeUsd": 0.00271,
    "gasUsd": 0.058491,
    "netVsHoldingUsd": -0.141952,
    "beatHolding": false
  },
  {
    "cycle": 4,
    "tokenId": "7319374",
    "mintIso": "2026-09-03T16:02:41.000Z",
    "closeIso": "2026-09-05T08:24:52.000Z",
    "heldSeconds": 145331,
    "entryNotionalUsd": 39.605351,
    "bnbMovePercent": 3.0683,
    "impermanentLossUsd": -0.09716,
    "swapFeesUsd": 0.055877,
    "cakeUsd": 0.012331,
    "gasUsd": 0.076047,
    "netVsHoldingUsd": -0.104999,
    "beatHolding": false
  },
  {
    "cycle": 5,
    "tokenId": "7336818",
    "mintIso": "2026-09-05T08:25:01.000Z",
    "closeIso": "2026-09-05T14:55:31.000Z",
    "heldSeconds": 23430,
    "entryNotionalUsd": 40.003445,
    "bnbMovePercent": 2.8071,
    "impermanentLossUsd": -0.085796,
    "swapFeesUsd": 0.028454,
    "cakeUsd": 0.002956,
    "gasUsd": 0.067806,
    "netVsHoldingUsd": -0.122192,
    "beatHolding": false
  }
] as const;

/** The six positions, from E2 winRate A (time in range). */
export const POSITIONS = [
  {
    "tokenId": "7268711",
    "mintBlock": 118536451,
    "closeBlock": 118625020,
    "heldSeconds": 39866,
    "inRangeFraction": 1,
    "rangeExits": 0,
    "smallestHeadroomTicks": 313
  },
  {
    "tokenId": "7272152",
    "mintBlock": 118625047,
    "closeBlock": 119714487,
    "heldSeconds": 490633,
    "inRangeFraction": 1,
    "rangeExits": 0,
    "smallestHeadroomTicks": 367
  },
  {
    "tokenId": "7316556",
    "mintBlock": 119714506,
    "closeBlock": 119754973,
    "heldSeconds": 18217,
    "inRangeFraction": 1,
    "rangeExits": 0,
    "smallestHeadroomTicks": 336
  },
  {
    "tokenId": "7319374",
    "mintBlock": 119755292,
    "closeBlock": 120078128,
    "heldSeconds": 145331,
    "inRangeFraction": 1,
    "rangeExits": 0,
    "smallestHeadroomTicks": 348
  },
  {
    "tokenId": "7336818",
    "mintBlock": 120078147,
    "closeBlock": 120130205,
    "heldSeconds": 23430,
    "inRangeFraction": 1,
    "rangeExits": 0,
    "smallestHeadroomTicks": 368
  },
  {
    "tokenId": "7340903",
    "mintBlock": 120130230,
    "closeBlock": null,
    "heldSeconds": 67033,
    "inRangeFraction": 1,
    "rangeExits": 0,
    "smallestHeadroomTicks": 510
  }
] as const;

/** The nine-day economics, numeric, from E2 gasVersusIncome / yields / netVersusHolding. */
export const ECON = {
  "windowDays": 9.0821,
  "windowStartIso": "2026-08-28T07:34:41.000Z",
  "windowEndIso": "2026-09-06T09:32:56.000Z",
  "startBlock": 118536451,
  "endBlock": 120279139,
  "transactions": 48,
  "incomeUsd": 0.271942,
  "swapFeesUsd": 0.211725,
  "cakeUsd": 0.060218,
  "gasUsd": 0.368063,
  "grossYieldPct": 0.6886,
  "twaNotionalUsd": 39.490415,
  "breakevenUsd": 53.45,
  "entryUsd": 39.958556,
  "agentNetUsd": 39.845391,
  "holdEntryBasketUsd": 41.771576,
  "netVsHoldPct": -4.6112,
  "observabilityGapFromIso": "2026-08-28T18:39:13.000Z",
  "observabilityGapToIso": "2026-09-03T10:56:32.000Z"
} as const;